% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_application.R
\name{approx_autocor}
\alias{approx_autocor}
\title{Function to Approximate an Autocorrelation Function Based on a Surface}
\usage{
approx_autocor(cov_dat, lag = 100)
}
\arguments{
\item{cov_dat}{Data set containing the surface information as given by
covariance_surf_plot_helper().}

\item{lag}{Maximal lag possible. Default is 100 such as the evaluation points
of the surface.}
}
\description{
This function takes a data set given by the covariance_surf_plot_helper()
function and approximates an autocorrelation function by taking the mean of
the off-diagonals.
}
