% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{cov2DataFrame}
\alias{cov2DataFrame}
\title{Compute the Surface of a Covariance Compoent and Convert it to Data.Frame}
\usage{
cov2DataFrame(m_aco, m_epg, component)
}
\arguments{
\item{m_aco}{Univariate model on dimension aco.}

\item{m_epg}{Univariate model on dimension epg.}

\item{component}{Covariance component to extract}
}
\description{
Compute the Surface of a Covariance Compoent and Convert it to Data.Frame
}
