% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{cova_comp_plot}
\alias{cova_comp_plot}
\title{Plot the Comparison of Univariate and Multivariate Covariate Effects}
\usage{
cova_comp_plot(m_true_comp, aco_pr, epg_pr, effects = c(4, 5, 6),
  effects_uni = c(5, 6, 7), m_fac = 2, limits = limits)
}
\arguments{
\item{m_true_comp}{Model components of the multivariate model.}

\item{aco_pr}{Predictions for aco model.}

\item{epg_pr}{Predictions for epg model.}

\item{effects}{Index number of the effects to be plotted from multivariate
model.}

\item{effects_uni}{Index number of the effects to be plotted from the
univariate}

\item{m_fac}{Multiplication factor.}

\item{limits}{Y-limits of the Plot.}
}
\description{
Plot the Comparison of Univariate and Multivariate Covariate Effects
}
