% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_plots.R
\name{covariance_surf_plot_helper}
\alias{covariance_surf_plot_helper}
\title{Extract Data for Covariance Surface Plot}
\usage{
covariance_surf_plot_helper(mcomp, component, dimlabels, corr = FALSE)
}
\arguments{
\item{mcomp}{The extracted model components as given by the internal function
extract_components().}

\item{component}{Which FPC component to extract. Possible values ("B", "C",
"E")}

\item{dimlabels}{The dimensions of the model.}

\item{corr}{TRUE if the correlation surface is to be estimated. Defaults to
FALSE. Only implemented for two dimensions.}
}
\description{
This function gives as an output the data set necessary to
plot the covariance surface. Probably not apt for more than two dimensions.
}
