% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{covariate_plot}
\alias{covariate_plot}
\title{Plot Covariate Effects Together}
\usage{
covariate_plot(m_true_comp, effects = c(4, 5, 6), m_fac = 2,
  limits = limits, dimlabels = c("ACO", "EPG"))
}
\arguments{
\item{m_true_comp}{True model components.}

\item{effects}{Number of effect function to be plotted.}

\item{m_fac}{Multiplication factor.}

\item{limits}{Y-limits for the plot.}

\item{dimlabels}{Labels for the dimensions in the plot.}
}
\description{
Plot Covariate Effects Together
}
