% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{covSurv}
\alias{covSurv}
\title{Estimate Covariance Surface}
\usage{
covSurv(vals, fcts, dim1, dim2, multi = TRUE)
}
\arguments{
\item{vals}{Eigenvalues (Vector).}

\item{fcts}{Eigenfunctions (MultiFunData).}

\item{dim1}{Choose first dimension (Scalar).}

\item{dim2}{Choose second dimension (Scalar).}

\item{multi}{Is the input multivariate (if not then vals and fcts are lists).}
}
\description{
Estimate Covariance Surface
}
