% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_application.R
\name{estim_overall_cov}
\alias{estim_overall_cov}
\title{Function to Estimate the Covariance Operator}
\usage{
estim_overall_cov(data, m_mean = c(2, 1), covariate = FALSE,
  num_covariates = 4, covariate_form = rep("by", times = 4),
  interaction = FALSE, which_interaction = matrix(NA))
}
\arguments{
\item{data}{Data.table containig the information needed for the sparseflmm()
function}
}
\description{
This function conducts a fast symmetric covariance estimation on some
data.table as proposed by Cederbaum et al. (2018). The parameters in the
function definition depend on the implementation of the sparseflmm(). Given
the estimated covariance surface, the function also conducts an MFPCA thus
giving FPCs.
}
