% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{extract_Covfct_sim_uni}
\alias{extract_Covfct_sim_uni}
\title{Extract the Estimated Covariate Effect for One Specific Model Term of the
Univariate Estimation}
\usage{
extract_Covfct_sim_uni(term, term_uni, m_true_comp, cov_preds)
}
\arguments{
\item{term}{Name of covariate effect term.}

\item{term_uni}{Name of covariate in univariate estimation ("int",
"famm_cb_mean", "famm_cb_covariate.1", "famm_cb_inter_1_2").}

\item{m_true_comp}{True model components.}

\item{cov_preds}{Simulation results of covariate effects.}
}
\description{
Extract the Estimated Covariate Effect for One Specific Model Term of the
Univariate Estimation
}
