% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{extract_Eigenfct_sim_dim}
\alias{extract_Eigenfct_sim_dim}
\title{Extract the Estimated Eigenfunctions for One Specific Variance Component for
a different number of estimated fcts}
\usage{
extract_Eigenfct_sim_dim(number, component, m_true_comp, eigenfcts,
  flip = TRUE)
}
\arguments{
\item{number}{Number of eigenfunction to be plotted.}

\item{component}{Name of variance component.}

\item{m_true_comp}{True model component.}

\item{eigenfcts}{Simulation results of eigenfcts.}

\item{flip}{Are the eigenfunctions to be flipped?}
}
\description{
Extract the Estimated Eigenfunctions for One Specific Variance Component for
a different number of estimated fcts
}
