% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{extract_Eigenfct_sim_uni}
\alias{extract_Eigenfct_sim_uni}
\title{Extract the Estimated Eigenfunctions for One Specific Variance Component of
the Univariate Estimation}
\usage{
extract_Eigenfct_sim_uni(number, component, eigenfcts)
}
\arguments{
\item{number}{Number of eigenfunction to be plotted.}

\item{component}{Name of variance component.}

\item{eigenfcts}{Simulation results of eigenfcts that have been filled up.}
}
\description{
Extract the Estimated Eigenfunctions for One Specific Variance Component of
the Univariate Estimation
}
