% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{flip_scores_uni}
\alias{flip_scores_uni}
\title{Flip Estimated Eigenscores According to the Flipping of Univariate
Eigenfunctions}
\usage{
flip_scores_uni(fun_true, fun_estim, score_estim)
}
\arguments{
\item{fun_true}{True eigenfunctions as a funData object.}

\item{fun_estim}{Estimated eigenfunctions as a funData object.}

\item{score_estim}{Estimated scores.}
}
\description{
Flip Estimated Eigenscores According to the Flipping of Univariate
Eigenfunctions
}
