% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{fpc2DataFrame}
\alias{fpc2DataFrame}
\title{Convert Univariate Estimates of Eigenfunctions to Data.frame}
\usage{
fpc2DataFrame(phi_aco, phi_epg, grid = seq(0, 1, length.out = 100))
}
\arguments{
\item{phi_aco}{Univariate eigenfunction for model aco.}

\item{phi_epg}{Univariate eigenfunction for model epg.}

\item{grid}{Grid of evaluation points.}
}
\description{
Convert Univariate Estimates of Eigenfunctions to Data.frame
}
