% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_application.R
\name{plot_snooker_fit}
\alias{plot_snooker_fit}
\title{Plot the Snooker Fit per Observation}
\usage{
plot_snooker_fit(obs, model, model_comp = NULL)
}
\arguments{
\item{obs}{Named level of the observation to be plotted.}

\item{model}{Model to be evaluated.}

\item{model_comp}{Additional model if two models are to be compared.}
}
\description{
This function plots the observed two-dimensional snooker trajectories against
their fitted trajectories.
}
